+++ var/bin/build-ddev-installation.sh
@@ -1,15 +1,48 @@
-#!/bin/bash
+#!/usr/bin/env bash
+################################################################################
+#
+#  This script install bootstrap_italia theme for local development
+#
+#  Author: bootstrap_italia maintainers
+#  Site: https://drupal.org/project/bootstrap_italia
+#  License: GPL-2.0-only
+#
+#  DDEV docs: https://ddev.readthedocs.io/en/latest/#intro-to-ddev-local
+#
+################################################################################
 
-# https://ddev.readthedocs.io/en/latest/#intro-to-ddev-local
+# Settings
+drupal_versions="10|11"
+default_drupal_version="10"
+bootstrap_italia_versions="2.x|2.x-dev@dev|latest"
+default_bootstrap_italia_version="2.x-dev@dev"
+vanilla_library="https://github.com/italia/bootstrap-italia/releases/download/v2.7.0/bootstrap-italia.zip"
 
+# Functions
+validate_project_name() {
+  local input=$1
+  local fqdn_host_regex='^[a-z][a-z0-9_-]*[a-z0-9]$'
+
+  if [[ ! $input =~ $fqdn_host_regex ]]; then
+      echo "The project name is not in the correct format."
+      return 1
+  fi
+}
+
+# Script
 echo "==[ Settings ]=="
 
 read -r -p "Project machine name [a-zA-Z0-9] (new-project): " project_name
 project_name=${project_name:-new-project}
 
+while ! validate_project_name "$project_name"; do
+  read -r -p "Project machine name [a-zA-Z0-9] (new-project): " project_name
+  project_name=${project_name:-new-project}
+done
+
 if [ -d "$project_name" ]; then
   echo "${project_name} already exists! Exit..."
-  exit
+  exit 1
 fi
 
 # Search project in ddev
@@ -19,15 +52,30 @@ if [ "$project_name" == "$ddev_search" ]; then
   exit
 fi
 
-read -r -p "Drupal version [9|10] (9): " drupal_version
-if [ -z "$drupal_version" ] || [ "$drupal_version" == "9" ]; then
-  drupal_version="9"
-else
-  drupal_version="10"
+while true; do
+    read -r -p "Drupal version [$drupal_versions] ($default_drupal_version): " drupal_version
+
+    # Set the default value if the input is empty
+    drupal_version=${drupal_version:-$default_drupal_version}
+
+    # Check if the input is correct
+    if [[ "$drupal_version" =~ ^($drupal_versions)$ ]]; then
+        break
+    else
+        echo "Error: Please enter only '10' or '11'. Try again."
+    fi
+done
+
+if [ "$drupal_version" != "$default_drupal_version" ]; then
+  read -r -p "The $drupal_version is unstable, use it for development only. Do you really want to install $drupal_version (if you choose no, $default_drupal_version will be installed)? [Y/n]: " unstable_response
+  if [[ "$unstable_response" =~ ^[Nn]$ ]]; then
+    drupal_version=$default_drupal_version
+    echo "Install version = ${default_drupal_version}."
+  fi
 fi
 
-read -r -p "Bootstrap Italia version [2.x|2.x-dev@dev|latest] (2.x-dev@dev): " bootstrap_italia_version
-bootstrap_italia_version=${bootstrap_italia_version:-2.x-dev@dev}
+read -r -p "Bootstrap Italia version [$bootstrap_italia_versions] ($default_bootstrap_italia_version): " bootstrap_italia_version
+bootstrap_italia_version=${bootstrap_italia_version:-$default_bootstrap_italia_version}
 
 read -r -p "Do you want enable italian language? [y|n] (y): " enable_locale
 enable_locale=${enable_locale:-y}
@@ -47,14 +95,23 @@ enable_experimental_modules=${enable_experimental_modules:-n}
 read -r -p "Bootstrap libraries type [vanilla|webpack] (webpack): " bi_libraries_type
 bi_libraries_type=${bi_libraries_type:-webpack}
 
-read -r -p "Do you want enable dev mode [y|n] (n): " enable_dev_mode
+read -r -p "Do you want enable dev mode? [y|n] (n): " enable_dev_mode
 enable_dev_mode=${enable_dev_mode:-n}
+
+# Admin Password
+read -r -p "Enter the admin password [minimum 12 characters] (random): " admin_password
+while [[ -z $admin_password || ${#admin_password} -lt 12 ]]; do
+  admin_password=$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 12 | head -n 1)
+done
 
 echo "==[ Configuration ]=="
 
 echo "Make ${project_name}"
 mkdir "$project_name"
-cd "$project_name" || exit
+cd "$project_name" || {
+    echo "Unable to change into directory $project_name"
+    exit 1
+}
 
 echo "Installing ddev config files in $(pwd) ..."
 ddev config --project-type=drupal${drupal_version} --docroot=web --create-docroot
@@ -63,17 +120,12 @@ echo 'Run docker containers'
 ddev start
 
 echo "Download Drupal ${drupal_version} and drush"
-if [ "$drupal_version" == "9" ]; then
-  ddev composer create 'drupal/recommended-project:^9' --no-install
-else
-  ddev composer create --no-install 'drupal/recommended-project:^10'
-fi
-
+ddev composer create --no-install "drupal/recommended-project:^$drupal_version"
 ddev composer require drush/drush --no-install
 ddev composer install
 
 echo "==[ Installing Drupal ${drupal_version} ]=="
-ddev exec drush -y site:install
+ddev exec drush -y site:install "--account-pass=${admin_password}"
 
 echo '==[ Install end enable bootstrap_italia dependencies ]=='
 ddev exec drush -y pm:enable inline_form_errors responsive_image
@@ -132,8 +184,7 @@ if [ "$bi_libraries_type" == "vanilla" ]; then
     mkdir ./web/themes/custom/italiagov/dist
   fi
 
-  curl --request GET -sL \
-    --url 'https://github.com/italia/bootstrap-italia/releases/download/v2.7.0/bootstrap-italia.zip' \
+  curl --request GET -sL --url $vanilla_library \
     --output './web/themes/custom/italiagov/dist/bootstrap-italia.zip'
 
   unzip ./web/themes/custom/italiagov/dist/bootstrap-italia.zip -d ./web/themes/custom/italiagov/dist/
@@ -154,13 +205,10 @@ if [ "$enable_modules" == "y" ]; then
   fi
 
   echo 'Install module: Bootstrap Italia Image Styles'
-  if [ "$drupal_version" == "10" ]; then
-    ddev composer require 'drupal/focal_point:^2.0@alpha'
-  else
-    ddev composer require drupal/focal_point
-  fi
+  ddev composer require drupal/focal_point
   ddev exec drush -y pm:enable responsive_image focal_point bootstrap_italia_image_style
 
+  # Deprecated
   if [ "$drupal_version" == "9" ]; then
     echo 'Install module: Bootstrap Italia Text Editor'
     ddev exec drush -y pm:enable bootstrap_italia_text_editor
@@ -223,11 +271,8 @@ if [ "$enable_modules" == "y" ]; then
   ddev exec drush -y pm:enable bootstrap_italia_paragraph_timeline
 
   echo 'Install module: Bootstrap Italia Paragraph Webform'
-  if [ "$drupal_version" == "10" ]; then
-    ddev composer require 'drupal/webform:^6.2@beta'
-  else
-    ddev composer require drupal/webform
-  fi
+  ddev composer require 'drupal/webform:^6.2@beta'
+
   ddev composer require wikimedia/composer-merge-plugin
   ddev exec drush -y pm:enable webform webform_bootstrap webform_ui bootstrap_italia_paragraph_webform
   ddev exec sed "-i 's#\"extra\": {#\"extra\": {\n\ \ \ \ \ \ \ \ \"merge-plugin\":{ \"include\": [\"web/modules/contrib/webform/composer.libraries.json\"] },#g' /var/www/html/composer.json"
@@ -253,17 +298,15 @@ if [ "$enable_experimental_modules" == "y" ]; then
   echo "==[ Install experimental modules ]=="
 
   ddev exec drush -y pm:enable ckeditor5 bootstrap_italia_text_editor2
-
 fi
 
 if [ "$enable_dev_mode" == "y" ]; then
   echo "==[ Enable dev mode ]=="
   ddev composer require drupal/devel drupal/dev_mode
   ddev exec drush -y pm:enable devel dev_mode
-
 fi
 
-if [[ "$enable_content_type" == "y" && "$drupal_version" == "9" ]]; then
+if [[ "$enable_content_type" == "y" ]]; then
   echo "==[ Install content type ]=="
 
   echo 'Install module: Bootstrap Italia content News'
@@ -289,11 +332,7 @@ fi
 if [ "$enable_ui_tools" == "y" ]; then
   echo "==[ Install UI tools ]=="
 
-  if [ "$drupal_version" == "10" ]; then
-    ddev composer require drupal/ui_patterns 'drupal/ui_patterns_settings:^2.0@RC' drupal/ds
-  else
-    ddev composer require drupal/ui_patterns drupal/ui_patterns_settings drupal/ds
-  fi
+  ddev composer require drupal/ui_patterns drupal/ui_patterns_settings drupal/ds
   ddev exec drush -y pm:enable ui_patterns ui_patterns_settings \
     layout_discovery layout_builder bootstrap_italia_layouts \
     ds ds_extras ds_switch_view_mode \
@@ -306,13 +345,13 @@ ddev auth ssh
 # ddev info
 ddev describe
 
-# Reset password message
-echo -e "Run -> \e[1m\e[31m ddev exec drush upwd admin my_password \e[0m <- if you don't know the Drupal administrator password"
+# Admin password
+echo "Admin password: ${admin_password}"
+echo -e "Run -> \e[1m\e[31m ddev exec drush upwd admin my_password \e[0m <- to choose new password"
 
 # Open browser
 #ddev launch
 
 # Remove test
 # ddev stop $project_name && ddev delete $project_name
-
 # then delete $project_name folder